reinstalled Next app
TypeScript - Yes
ESLint - Yes
Tailwind CSS - No
src/` directory - Yes
App Router - Yes
customise the 'import alias' - No
