+++ package.json
@@ -9,13 +9,14 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "autoprefixer": "10.4.16",
+    "@types/node": "20.8.10",
+    "@types/react": "18.2.33",
+    "@types/react-dom": "18.2.14",
     "eslint": "8.52.0",
-    "eslint-config-next": "14.0.0",
-    "next": "14.0.0",
-    "postcss": "8.4.31",
+    "eslint-config-next": "14.0.1",
+    "next": "14.0.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "tailwindcss": "3.3.5"
+    "typescript": "5.2.2"
   }
 }
